# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Papers Please System API
  description: API для системы управления УПК (пограничные пункты)
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearerAuth: []
paths:
...
  /api/v1/upks:
    get:
      tags:
      - upk-controller
      operationId: getAllUpks
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedResponseUpkResponse"
    post:
      tags:
      - upk-controller
      operationId: createUpk
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpkRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UpkResponse"
  /api/v1/tickets:
    get:
      tags:
      - ticket-controller
      operationId: getAllTickets
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedResponseTicketResponse"
    post:
      tags:
      - ticket-controller
      operationId: createTicket
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TicketResponse"
  /api/v1/tickets/{id}/reopen:
    post:
      tags:
      - ticket-controller
      operationId: reopenTicket
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TicketResponse"
  /api/v1/tickets/{id}/close:
    post:
      tags:
      - ticket-controller
      operationId: closeTicket
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: resolution
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TicketResponse"
  /api/v1/tickets/{id}/assign:
    post:
      tags:
      - ticket-controller
      operationId: assignExecutor
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: executorId
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TicketResponse"
...
  /api/v1/upks/{id}:
    get:
      tags:
      - upk-controller
      operationId: getUpkById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UpkResponse"
    delete:
      tags:
      - upk-controller
      operationId: deleteUpk
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Unit"
    patch:
      tags:
      - upk-controller
      operationId: partialUpdateUpk
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpkRequestPartial"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UpkResponse"
  /api/v1/tickets/{id}:
    get:
      tags:
      - ticket-controller
      operationId: getTicketById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TicketResponse"
    delete:
      tags:
      - ticket-controller
      operationId: deleteTicket
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Unit"
    patch:
      tags:
      - ticket-controller
      operationId: partialUpdateTicket
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketRequestPartial"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TicketResponse"
...
  /api/v1/tickets/by-type/{type}:
    get:
      tags:
      - ticket-controller
      operationId: getTicketsByType
      parameters:
      - name: type
        in: path
        required: true
        schema:
          type: string
          enum:
          - EXTERNAL
          - INTERNAL
          - ARREST
          - CROSSCHECK
          - APPEAL
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedResponseTicketResponse"
  /api/v1/tickets/by-status/{status}:
    get:
      tags:
      - ticket-controller
      operationId: getTicketsByStatus
      parameters:
      - name: status
        in: path
        required: true
        schema:
          type: string
          enum:
          - OPEN
          - IN_PROGRESS
          - NEED_INFO
          - CLOSED
          - REJECTED
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedResponseTicketResponse"
  /api/v1/tickets/by-shift/{shiftId}:
    get:
      tags:
      - ticket-controller
      operationId: getTicketsByShift
      parameters:
      - name: shiftId
        in: path
        required: true
        schema:
          type: string
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedResponseTicketResponse"
  /api/v1/tickets/by-priority/{priority}:
    get:
      tags:
      - ticket-controller
      operationId: getTicketsByPriority
      parameters:
      - name: priority
        in: path
        required: true
        schema:
          type: string
          enum:
          - LOW
          - NORMAL
          - HIGH
          - CRITICAL
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedResponseTicketResponse"
  /api/v1/tickets/by-executor/{executorId}:
    get:
      tags:
      - ticket-controller
      operationId: getTicketsByExecutor
      parameters:
      - name: executorId
        in: path
        required: true
        schema:
          type: string
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedResponseTicketResponse"
  /api/v1/tickets/by-author/{authorId}:
    get:
      tags:
      - ticket-controller
      operationId: getTicketsByAuthor
      parameters:
      - name: authorId
        in: path
        required: true
        schema:
          type: string
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedResponseTicketResponse"
...
components:
  schemas:
...
    UpkRequest:
      required:
      - bossId
      - name
      - region
      type: object
      properties:
        name:
          maxLength: 100
          minLength: 2
          type: string
        bossId:
          type: string
        region:
          type: string
          enum:
          - ORVECH_VONOR
          - EAST_GRESTIN
          - PARADIZNA
    UpkResponse:
      required:
      - bossId
      - id
      - name
      - region
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        bossId:
          type: string
        region:
          type: string
          enum:
          - ORVECH_VONOR
          - EAST_GRESTIN
          - PARADIZNA
    TicketRequest:
      required:
      - authorId
      - priority
      - status
      - ticketType
      type: object
      properties:
        ticketType:
          type: string
          enum:
          - EXTERNAL
          - INTERNAL
          - ARREST
          - CROSSCHECK
          - APPEAL
        status:
          type: string
          enum:
          - OPEN
          - IN_PROGRESS
          - NEED_INFO
          - CLOSED
          - REJECTED
        priority:
          type: string
          enum:
          - LOW
          - NORMAL
          - HIGH
          - CRITICAL
        deadlineAt:
          type: string
          format: date-time
        authorId:
          type: string
        executorId:
          type: string
        parentTicketId:
          type: string
    TicketResponse:
      required:
      - authorId
      - createdAt
      - id
      - priority
      - status
      - ticketType
      - updatedAt
      type: object
      properties:
        id:
          type: string
        ticketType:
          type: string
          enum:
          - EXTERNAL
          - INTERNAL
          - ARREST
          - CROSSCHECK
          - APPEAL
        status:
          type: string
          enum:
          - OPEN
          - IN_PROGRESS
          - NEED_INFO
          - CLOSED
          - REJECTED
        priority:
          type: string
          enum:
          - LOW
          - NORMAL
          - HIGH
          - CRITICAL
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deadlineAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
        authorId:
          type: string
        executorId:
          type: string
        parentTicketId:
          type: string
...
    UpkRequestPartial:
      type: object
      properties:
        name:
          maxLength: 100
          minLength: 2
          type: string
        bossId:
          type: string
        region:
          type: string
          enum:
          - ORVECH_VONOR
          - EAST_GRESTIN
          - PARADIZNA
    TicketRequestPartial:
      type: object
      properties:
        ticketType:
          type: string
          enum:
          - EXTERNAL
          - INTERNAL
          - ARREST
          - CROSSCHECK
          - APPEAL
        status:
          type: string
          enum:
          - OPEN
          - IN_PROGRESS
          - NEED_INFO
          - CLOSED
          - REJECTED
        priority:
          type: string
          enum:
          - LOW
          - NORMAL
          - HIGH
          - CRITICAL
        deadlineAt:
          type: string
          format: date-time
        authorId:
          type: string
        executorId:
          type: string
        resolvedAt:
          type: string
          format: date-time
...
    PagedResponseUpkResponse:
      required:
      - items
      - limit
      - offset
      - total
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/UpkResponse"
        total:
          type: integer
          format: int64
        limit:
          type: integer
          format: int32
        offset:
          type: integer
          format: int32
    PagedResponseTicketResponse:
      required:
      - items
      - limit
      - offset
      - total
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/TicketResponse"
        total:
          type: integer
          format: int64
        limit:
          type: integer
          format: int32
        offset:
          type: integer
          format: int32
...
    Unit:
      type: object
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT